tests/wpt/web-platform-tests/eventsource/eventsource-constructor-url-bogus.htm

<!DOCTYPE html>
<html>
  <head>
    <title>EventSource: constructor (invalid URL)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        assert_throws("SyntaxError", function() { new EventSource("http://this is invalid/") })
      })
    </script>
  </body>
</html>
